move result object to a different file
