+++ alphastats/gui/utils/ui_helper.py
@@ -154,3 +154,23 @@ class StateKeys(metaclass=ConstantsClass):
     SELECTED_UNIPROT_FIELDS = "selected_uniprot_fields"
 
     ORGANISM = "organism"  # TODO this is essentially a constant
+
+
+class ANALYSIS_PARAMETERS(metaclass=ConstantsClass):
+    TWOGROUP_GROUP1 = "group1"
+    TWOGROUP_GROUP2 = "group2"
+    DEA_TWOGROUPS_METHOD = "method"
+    DEA_TWOGROUPS_FDR_METHOD = "fdr_method"
+    TWOGROUP_COLUMN = "column"
+
+
+class RESULT_PARAMETERS(metaclass=ConstantsClass):
+    WIDTH = "width"
+    HEIGHT = "height"
+    SHOWLEGEND = "showlegend"
+    QVALUE_CUTOFF = "qvalue_cutoff"
+    LOG2FC_CUTOFF = "log2fc_cutoff"
+    FLIP_XAXIS = "flip_xaxis"
+    DRAWLINES = "drawlines"
+    LABEL_SIGNIFICANT = "label_significant"
+    RENDERER = "renderer"
